refactor(auth): consolidate env-management commands into manage.py
list / default / status / logout all load the config file and act on a named
environment, and duplicated the config-load block (x4) and the interactive env
picker (x2). Merge the four into cli/login/manage.py with local _load_config and
_select_environment helpers, and move the display helpers (environment_summary,
print_environments, describe_scope, describe_token_cache) there from _common,
since these commands are their only consumers.

_common.py is now single-purpose (login-parameter resolution); the login package
drops from 11 modules to 8 and ~48 lines. Behavior and --help output unchanged.
