+++ tests/bfabric_scripts/cli/login/test_cmd_auth_default.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import yaml
 
-from bfabric_scripts.cli.login.default_config import cmd_auth_default
+from bfabric_scripts.cli.login.manage import cmd_auth_default
 
 
 def _write_config(config_file, default="PROD"):
@@ -46,8 +46,8 @@ class TestInteractivePicker:
         config_file = tmp_path / "config.yml"
         _write_config(config_file, default="PROD")
         # The picker runs only when no value is passed; it returns the chosen environment.
-        mocker.patch("bfabric_scripts.cli.login.default_config.is_interactive", return_value=True)
-        picker = mocker.patch("bfabric_scripts.cli.login.default_config.select_choice", return_value="TEST")
+        mocker.patch("bfabric_scripts.cli.login.manage.is_interactive", return_value=True)
+        picker = mocker.patch("bfabric_scripts.cli.login.manage.select_choice", return_value="TEST")
         cmd_auth_default(config_file=config_file)
         data = yaml.safe_load(config_file.read_text())
         assert data["GENERAL"]["default_config"] == "TEST"
@@ -58,8 +58,8 @@ class TestInteractivePicker:
         config_file = tmp_path / "config.yml"
         _write_config(config_file, default="PROD")
         before = config_file.read_text()
-        mocker.patch("bfabric_scripts.cli.login.default_config.select_choice", return_value=None)
-        mocker.patch("bfabric_scripts.cli.login.default_config.is_interactive", return_value=True)
+        mocker.patch("bfabric_scripts.cli.login.manage.select_choice", return_value=None)
+        mocker.patch("bfabric_scripts.cli.login.manage.is_interactive", return_value=True)
         cmd_auth_default(config_file=config_file)
         assert "No changes made" in capsys.readouterr().out
         assert config_file.read_text() == before
@@ -70,7 +70,7 @@ class TestNonInteractiveListing:
         config_file = tmp_path / "config.yml"
         _write_config(config_file, default="PROD")
         before = config_file.read_text()
-        mocker.patch("bfabric_scripts.cli.login.default_config.is_interactive", return_value=False)
+        mocker.patch("bfabric_scripts.cli.login.manage.is_interactive", return_value=False)
         cmd_auth_default(config_file=config_file)
         output = capsys.readouterr().out
         assert "PROD" in output
